Move parser classes into own module
Helps clean up the concerns directory, since these classes
are no longer concerns.
