+++ test/models/dropbox_test.rb
@@ -27,22 +27,22 @@ class DropboxUnitTest <  ActiveSupport::TestCase
   end
 
   test "returns provider and type" do
-    assert_equal ParserDropboxItem.type, 'dropbox_item'
+    assert_equal Parser::DropboxItem.type, 'dropbox_item'
   end
 
   test "matches known URL patterns, and returns instance on success" do
-    assert_nil ParserDropboxItem.match?('https://example.com')
+    assert_nil Parser::DropboxItem.match?('https://example.com')
     
-    match_one = ParserDropboxItem.match?('https://www.dropbox.com/s/fake-url/example.txt')
-    assert_equal true, match_one.is_a?(ParserDropboxItem)
-    match_two = ParserDropboxItem.match?('https://dropboxusercontent.com/s/fake-url/example.txt')
-    assert_equal true, match_two.is_a?(ParserDropboxItem)
-    match_three = ParserDropboxItem.match?('https://www.dropbox.com/sh/fake-url/example.txt')
-    assert_equal true, match_three.is_a?(ParserDropboxItem)
+    match_one = Parser::DropboxItem.match?('https://www.dropbox.com/s/fake-url/example.txt')
+    assert_equal true, match_one.is_a?(Parser::DropboxItem)
+    match_two = Parser::DropboxItem.match?('https://dropboxusercontent.com/s/fake-url/example.txt')
+    assert_equal true, match_two.is_a?(Parser::DropboxItem)
+    match_three = Parser::DropboxItem.match?('https://www.dropbox.com/sh/fake-url/example.txt')
+    assert_equal true, match_three.is_a?(Parser::DropboxItem)
   end
 
   test "assigns values to hash from the HTML doc" do
-    data = ParserDropboxItem.new('https://www.dropbox.com/sh/fake-url/example.txt').parse_data(doc)
+    data = Parser::DropboxItem.new('https://www.dropbox.com/sh/fake-url/example.txt').parse_data(doc)
     assert_equal 'How to use the Public folder.rtf', data[:title]
     assert_equal 'This was totally shared with Dropbox', data[:description]
     assert_equal 'https://www.dropbox.com/static/metaserver/static/images/spectrum-icons/generated/content/content-docx-large.png', data[:picture]
@@ -51,7 +51,7 @@ class DropboxUnitTest <  ActiveSupport::TestCase
   test "should fall back to fetching information from URL" do
     empty_doc = Nokogiri::HTML('')
 
-    data = ParserDropboxItem.new('https://www.dropbox.com/sh/fake-url/How%20to%20use%20the%20Public%20folder.rtf').parse_data(empty_doc)
+    data = Parser::DropboxItem.new('https://www.dropbox.com/sh/fake-url/How%20to%20use%20the%20Public%20folder.rtf').parse_data(empty_doc)
     assert_equal 'How to use the Public folder.rtf', data[:title]
     assert_equal 'Shared with Dropbox', data[:description]
   end
@@ -60,11 +60,11 @@ class DropboxUnitTest <  ActiveSupport::TestCase
     mocked_airbrake = MiniTest::Mock.new
     mocked_airbrake.expect :call, :return_value, [StandardError, Hash]
 
-    ParserDropboxItem.any_instance.stubs(:get_html_metadata).raises(NoMethodError.new("Faking in test"))
+    Parser::DropboxItem.any_instance.stubs(:get_html_metadata).raises(NoMethodError.new("Faking in test"))
 
     data = nil
     PenderAirbrake.stub(:notify, mocked_airbrake) do
-      data = ParserDropboxItem.new('https://www.dropbox.com/sh/fake-url/example.txt').parse_data(doc)
+      data = Parser::DropboxItem.new('https://www.dropbox.com/sh/fake-url/example.txt').parse_data(doc)
     end
     mocked_airbrake.verify
     assert_equal 1, data.keys.count
